+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### Changed
 - Fix bug from #867 where the view config is temporarily invalid due to null values.
+- Separate out hooks to allow for arbitrary gene slicing.
+- Update AnnData loader to handle artbitrary gene slicing.
 
 ## [1.1.6](https://www.npmjs.com/package/vitessce/v/1.1.6) - 2021-03-05